+++ controllers/ValidController.py
@@ -11,7 +11,7 @@ import pandas as pd
 
 #  LOGGING
 
-logger = logging.getLogger("APPLICATION")
+logger = logging.getLogger("VALID_CONTROLLER")
 
 
 #  CLASS
@@ -22,10 +22,24 @@ class ValidController:
   
   def __init__(self, app_ref):
     self.app = app_ref
-    logger.info("[ValidController] initialised sucessfully.")
+    logger.info("initialised sucessfully.")
     
   
   #  METHODS
+  
+  def validate_col(self, 
+                   target_df: pd.DataFrame, 
+                   target_col: str) -> str:
+    
+    #  for index
+    if target_col.strip().lower() == "index":
+      return "index"
+    #  for columns
+    output = next((col for col in target_df.columns if col.strip().lower() == target_col.strip().lower()), None)
+    if output is None:
+      raise ValueError("target column is not found.")
+    return output.strip()
+  
 
   def validate_preview_df(self,
                           lb_text: str,
@@ -35,13 +49,13 @@ class ValidController:
     
     # if schema not found, unable to match, failed
     if expectation is None:
-        err_msg = f"[ValidController] {lb_text} is not in the target schema."
+        err_msg = f"{lb_text} is not in the target schema."
         logger.error(err_msg, exc_info=True)
         self.app.comp_fact.build_reminder_box("Error", err_msg)
         return False
     # if the col name list not matched, failed
     if expectation != list(target_state):
-      err_msg = (f"[ValidController] Unexpected column names.\n"
+      err_msg = (f"Unexpected column names.\n"
               f"Expectated {expectation} but got {list(target_state.columns)}.")
       logger.error(err_msg, exc_info=True)
       self.app.comp_fact.build_reminder_box("Error", err_msg)