+++ models/SQLConnector.py
@@ -7,7 +7,7 @@ from models.config.SQLCommands import SQL_COMMANDS
 
 #  LOGGING
 
-logger = logging.getLogger("APPLICATION")
+logger = logging.getLogger("SQL_CONNECTOR")
 
 
 #  CLASS
@@ -41,7 +41,7 @@ class SQLConnector:
           cursor = connection.cursor()
           return connection, cursor
       except Error as err:
-          logger.error(f"[SQLConnector] Connection error: {err}", exc_info=True)
+          logger.error(f"Connection error: {err}", exc_info=True)
           return None, None
 
  
@@ -50,7 +50,7 @@ class SQLConnector:
           cursor.close()
       if connection:
           connection.close()
-          logger.info("[SQLConnector] terminated MySQL connection.")
+          logger.info("terminated MySQL connection.")
           
 
   #  learnt: the fn as wrappper, bringing simplicity with sql code adopted
@@ -65,7 +65,7 @@ class SQLConnector:
         cursor.execute(sql)
         connection.commit()
     except Error as ex:
-        logger.error(f"[SQLConnector] Execution error: {ex}", exc_info=True)
+        logger.error(f"Execution error: {ex}", exc_info=True)
         raise
     #  disconnection
     finally:
@@ -82,7 +82,7 @@ class SQLConnector:
   def initialise_database(self, db_init=False):
     try:
         self.execute_db(SQL_COMMANDS["create_database"], db_initialised=db_init)
-        logger.info("[SQLConnector] Database is created.")
+        logger.info("Database is created.")
     except Exception as ex:
         logger.error(f"failed to create database - {ex}", exc_info=True)
 
@@ -93,7 +93,7 @@ class SQLConnector:
         self.execute_db(SQL_COMMANDS["create_table_users"], db_initialised=db_init)
         self.execute_db(SQL_COMMANDS["create_table_activities"], db_initialised=db_init)
         self.execute_db(SQL_COMMANDS["create_table_components"], db_initialised=db_init)
-        logger.info("[SQLConnector] Tables are created.")
+        logger.info("Tables are created.")
     except Exception as ex:
         logger.error(f"failed to create tables - {ex}", exc_info=True)
     
@@ -104,26 +104,26 @@ class SQLConnector:
   
   def drop_database(self, db_init=False):
       self.execute_db(SQL_COMMANDS["drop_database"], db_initialised=db_init)
-      logger.info("[SQLConnector] Database is dropped.")
+      logger.info("Database is dropped.")
       
       
   def drop_tables(self, table: str, db_init=True):
     
     table_r = re.sub(r'[^a-zA-Z]', "", table).lower()
     if table_r not in ["users", "activities", "components"]:
-        raise ValueError(f"[SQLConnector] \"{table_r}\" is not valid.")
+        raise ValueError(f"\"{table_r}\" is not valid.")
     
     if (table_r == "users"):
         self.execute_db(SQL_COMMANDS["drop_table_users"], db_initialised=db_init)
-        logger.info(f"[SQLConnector] drop_tables: \"{table_r}\" is dropped.")
+        logger.info(f"drop_tables: \"{table_r}\" is dropped.")
     elif (table_r == "activities"):
         self.execute_db(SQL_COMMANDS["drop_table_activities"], db_initialised=db_init)
-        logger.info(f"[SQLConnector] drop_tables: \"{table_r}\" is dropped.")
+        logger.info(f"drop_tables: \"{table_r}\" is dropped.")
     elif (table_r == "components"):
         self.execute_db(SQL_COMMANDS["drop_table_components"], db_initialised=db_init)
-        logger.info(f"[SQLConnector] drop_tables: \"{table_r}\" is dropped.")
+        logger.info(f"drop_tables: \"{table_r}\" is dropped.")
     else:
-        err_msg = "[SQLConnector] drop_tables: table name does not match."
+        err_msg = "drop_tables: table name does not match."
         logger.error(err_msg, exc_info=True)
         raise ValueError(err_msg)